# public/swagger.yaml
openapi: 3.0.0
info:
  title: Genouest Account Manager (My) API
  description: # todo
  version: 1.4.32
...
paths:
...
  /project:
    tags:
      - projects
    get:
      operationId: list-projects
      summary: List all projects
      responses:
        '200':
          $ref: '#/components/responses/200-SendProjectArray'
        '401':
          $ref: '#/components/responses/401-NotAuthorized'
        '404':
          $ref: '#/components/responses/404-UserNotFound'
    post:
      operationId: add-projects
      summary: Add a new project
      requestBody:
        $ref: '#/components/requestBodies/Project'
      responses:
        '200':
          $ref: '#/components/responses/200-SendConfirmation'
        '401':
          $ref: '#/components/responses/401-NotAuthorized'
        '403':
          $ref: '#/components/responses/403-InvalidInput'
        '404':
          $ref: '#/components/responses/404-NotFound'
    put:
      operationId: edit-projects
      summary: Edit project details
      requestBody:
        $ref: '#/components/requestBodies/Project'
      responses:
        '200':
          $ref: '#/components/responses/200-SendConfirmation'
        '401':
          $ref: '#/components/responses/401-NotAuthorized'
        '403':
          $ref: '#/components/responses/403-InvalidInput'
        '404':
          $ref: '#/components/responses/404-UserNotFound'
        '5XX':
          $ref: '#/components/responses/500-ServerError'
  /project/{id}:
    tags:
      - projects
    parameters:
      - $ref: '#/components/parameters/pid'
    get:
      operationId: get-projects
      summary: Get project details
      responses:
        '200':
          $ref: '#/components/responses/200-SendProject'
        '401':
          $ref: '#/components/responses/401-NotAuthorized'
        '403':
          $ref: '#/components/responses/403-InvalidInput'
        '404':
          $ref: '#/components/responses/404-UserNotFound'
    post:
      operationId: update-projects
      summary: Update project details
      requestBody:
        $ref: '#/components/requestBodies/Project'
      responses:
        '200':
          $ref: '#/components/responses/200-SendConfirmation'
        '401':
          $ref: '#/components/responses/401-NotAuthorized'
        '403':
          $ref: '#/components/responses/403-UnsanitaryId'
        '404':
          $ref: '#/components/responses/404-UserNotFound'
        '5XX':
          $ref: '#/components/responses/500-ServerError'
    delete:
      operationId: delete-projects
      summary: Delete a project
      responses:
        '200':
          $ref: '#/components/responses/200-SendConfirmation'
        '401':
          $ref: '#/components/responses/401-NotAuthorized'
        '403':
          $ref: '#/components/responses/403-UnsanitaryId'
        '404':
          $ref: '#/components/responses/404-UserNotFound'
        '5XX':
          $ref: '#/components/responses/500-ServerError'
  /project/{id}/extend:
    tags:
      - projects
    parameters:
      - $ref: '#/components/parameters/pid'
    get:
      operationId: extend-projects
      summary: Extend project duration
      responses:
        '200':
          description: Project duration extended
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  expiration:
                    type: number
        '401':
          $ref: '#/components/responses/401-NotAuthorized'
        '403':
          $ref: '#/components/responses/403-UnsanitaryId'
        '404':
          $ref: '#/components/responses/404-NotFound'
  /project/{id}/users:
    tags:
      - projects
    parameters:
      - $ref: '#/components/parameters/pid'
    get:
      operationId: getUsers-projects
      summary: Get users in a project
      responses:
        '200':
          $ref: '#/components/responses/200-SendUserArray'
        '401':
          $ref: '#/components/responses/401-NotAuthorized'
        '403':
          $ref: '#/components/responses/403-UnsanitaryId'
        '404':
          $ref: '#/components/responses/404-UserNotFound'
  /project/{id}/request:
    tags:
      - projects
    parameters:
      - $ref: '#/components/parameters/pid'
    post:
      operationId: request-projects
      summary: Request a project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                user:
                  type: string
                request:
                  type: string
      responses:
        '200':
          $ref: '#/components/responses/200-SendConfirmation'
        '401':
          $ref: '#/components/responses/401-NotAuthorized'
        '403':
          $ref: '#/components/responses/403-InvalidInput'
        '404':
          $ref: '#/components/responses/404-NotFound'
        '5XX':
          $ref: '#/components/responses/500-ServerError'
  /ask/project:
    tags:
      - projects
    post:
      operationId: askNew-projects
      summary: Ask for a new project
      requestBody:
        $ref: '#/components/requestBodies/Project'
      responses:
        '200':
          $ref: '#/components/responses/200-SendConfirmation'
        '401':
          $ref: '#/components/responses/401-NotAuthorized'
        '403':
          $ref: '#/components/responses/403-InvalidInput'
        '404':
          $ref: '#/components/responses/404-UserNotFound'
        '5XX':
          $ref: '#/components/responses/500-ServerError'
  /pending/project:
    tags:
      - projects
    get:
      operationId: list_pending-projects
      summary: List pending projects
      responses:
        '200':
          $ref: '#/components/responses/200-SendProjectArray'
        '401':
          $ref: '#/components/responses/401-NotAuthorized'
        '404':
          $ref: '#/components/responses/404-UserNotFound'
  /pending/project/{uuid}:
    tags:
      - projects
    parameters:
      - in: path
        name: uuid
        required: true
        schema:
          type: string
        description: Project uuid
    delete:
      operationId: delete_pending-projects
      summary: Delete a pending project
      responses:
        '200':
          $ref: '#/components/responses/200-SendConfirmation'
        '401':
          $ref: '#/components/responses/401-NotAuthorized'
        '404':
          $ref: '#/components/responses/404-UserNotFound'
        '5XX':
          $ref: '#/components/responses/500-ServerError'
...
components:
  parameters:
...
    pid:
      in: paths
      name: id
      required: true
      schema:
        type: string
      description: Project id (name)
...
  schemas:
    Message:
      type: object
      properties:
        message:
          type: string
      example:
        message: 'User session not found'
...
    User:
      type: object
      properties:
        first_name:
          type: string
        last_name:
          type: string
    Project:
      type: object
      properties:
        id:
          type: string
        expiration:
          type: number
...
  requestBodies:
...
    Project:
      description: An input Project object
      required: true
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Project'
...
  responses:
    200-SendConfirmation:
      description: OK
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Message'
    200-SendProject:
      description: OK - Send Project
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Project'
...
    200-SendUserArray:
      description: OK - Send Users
      content:
        application/json:
          schema:
            type: array
            items:
              $ref: '#/components/schemas/User'
    200-SendProjectArray:
      description: OK - Send Projects
      content:
        application/json:
          schema:
            type: array
            items:
              $ref: '#/components/schemas/Project'
...
    401-NotAuthorized:
      description: Authorization information is missing or invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Message'
    403-InvalidInput:
      description: The given input caused an error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Message'
    403-UnsanitaryId:
      description: Input id is not sanitized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Message'
...
    404-NotFound:
      description: The specified resource was not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Message'
    404-UserNotFound:
      description: A user with the specified uid was not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Message'
    500-ServerError:
      description: An unexpected server error has occurred
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Message'